Create sysGenerator class to initialize system.
The sysGenerator class initialises the solar system,
when the generator is initialized, and stores it in a member variable with
a unique pointer.

The getSystem function can then be called which
returns the unique pointer to the created system, using the move() functino.

This returned pointer should be assigned to a pSystem
unique pointer, and then one has a complete initialized solar system.

The evolveSystem method can then be called on the system to evolve it for an
arbitrary time t and with timestep dt.