Implement total energy calculation.
Since we are only intersted in the total energy before and after the simulation
this is a separate function getEnergy in the pSystem class. This simplifies things.
If we were interested in the energy at every step of the simualtion it could be calculated
in the acceleration and velocity/position update functions, this would also avoid having
to loop over the particles again, just to calculate the total energy. This seemed too complicated
for the intended use.