Implement InitialConditionGenerator template.
Two classes inherit the template the solarSysGenerator and
randomSysGenerator. Both return a unique pointer of a pSystem by
calling the generateInitialConditions method. Both classes initialize the
systems when the constructor is called. Could be better if they performed this
when the generateInitialConditions function is called, but since the solarSystem is small
this does not cause much overhed in the main, and when n is not specified
a system with 0 planets is created, i.e. it does not cause overhead either.